+++ Numerical_Methods_Lab_3/PolynomialMember.cs
@@ -8,10 +8,10 @@ namespace Numerical_Methods_Lab_3
 {
     internal class PolynomialMember
     {
-        public int Grade { get; set; }
+        public double Grade { get; set; }
         public double Сoefficient { get; set; }
         public PolynomialMember() {}
-        public PolynomialMember(double сoefficient, int grade)
+        public PolynomialMember(double сoefficient, double grade)
         {
             Grade = grade;
             Сoefficient = сoefficient;
